% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauda-docs.R
\name{cauda.dag}
\alias{cauda.dag}
\title{Learn and plot a causal DAG from a data frame}
\usage{
cauda.dag(df, highlight = NULL, verbose = TRUE)
}
\arguments{
\item{df}{A cleaned numeric data frame (output of \code{cauda.prep()}).}

\item{highlight}{Character. Column name to highlight in red (e.g. your target variable). Default is NULL.}

\item{verbose}{Logical. Print a summary of nodes and edges. Default is TRUE.}
}
\value{
Invisibly returns the learned \code{bn} object from \pkg{bnlearn}.
}
\description{
Uses the Hill-Climbing algorithm from the \pkg{bnlearn} package to learn a
Bayesian network structure from data and plots it as a directed acyclic graph (DAG).
}
\examples{
df_ready <- cauda.prep(mtcars, verbose = FALSE)
dag <- cauda.dag(df_ready, highlight = "mpg", verbose = FALSE)

}
